# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil do usuário retornado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter perfil do usuário logado",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Autentica o usuário usando credenciais básicas (Basic Auth) enviadas exclusivamente no cabeçalho HTTP Authorization, e retorna um token de acesso JWT.",
        "operationId": "AuthController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Login realizado com sucesso, retorna usuário e token de acesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Login do usuário",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token renovado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          }
        },
        "summary": "Refresh do access token",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/payment-methods": {
      "get": {
        "description": "Retorna uma lista paginada de formas de pagamento associados às credenciais do usuário autenticado.",
        "operationId": "PaymentMethodController_getPaymentMethods",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página para paginação",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Número de itens por página",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de formas de pagamento retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentMethodResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de requisição, como pageSize excedendo o limite."
          },
          "401": {
            "description": "Token de autenticação inválido ou ausente."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar formas de pagamento",
        "tags": [
          "PaymentMethod"
        ]
      }
    },
...
    "/api/v1/plans/stats": {
      "get": {
        "operationId": "StatsController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Estatísticas de uso e limites do plano retornadas com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtém estatísticas de uso e limites do plano do usuário logado",
        "tags": [
          "Plans & Usage"
        ]
      }
    },
    "/api/v1/plans": {
      "get": {
        "operationId": "StatsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de planos disponíveis retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista todos os planos disponíveis",
        "tags": [
          "Plans & Usage"
        ]
      }
    },
...
    "/api/v1/employees": {
      "get": {
        "operationId": "EmployeeController_get",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página atual",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Itens por página",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Busca por nome ou apelido (FUN_NOME, FUN_APELIDO)",
            "schema": {
              "example": "João",
              "type": "string"
            }
          },
          {
            "name": "situation",
            "required": false,
            "in": "query",
            "description": "Situação do funcionário (FUN_SITUACAO)",
            "schema": {
              "type": "string",
              "enum": [
                "A",
                "I"
              ]
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "description": "ID da loja (LOJ_CODIGO)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "functionId",
            "required": false,
            "in": "query",
            "description": "ID da função (FCA_CODIGO)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de funcionários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar funcionários",
        "tags": [
          "Employee"
        ]
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "operationId": "EmployeeController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do funcionário (FUN_CODIGO)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do funcionário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhe do funcionário",
        "tags": [
          "Employee"
        ]
      }
    },
...
  },
  "info": {
    "title": "InfoBrasil API",
    "description": "API de integração do sistema Retaguarda (by InfoBrasil) com outros sistemas, como ERPs, e-commerce, etc. <br><br>\n      Se deseja solicitar um acesso, preencha o formulário em: <a href='https://info-api.infobrasilsistemas.com.br/integration/form'>Solicitar Acesso</a>\n      <br><br>\n      **O sistema conta com duas formas de autenticação:**<br> \n      - Basic Auth para login e obtenção de token JWT<br>\n      - JWT para acesso aos endpoints protegidos.\n      <br><br>\n      Para verificar a disponibilidade da API acesse nossa <a href='/status'>Página de Status</a>\n      <br><br>\n      A API é organizada em módulos, cada um responsável por uma área específica do sistema, como produtos, clientes, vendas, etc.",
    "version": "1.15.1",
    "contact": {
      "name": "InfoBrasil Sistemas",
      "url": "https://www.infobrasilsistemas.com.br",
      "email": "suporte@infobrasilsistemas.com.br"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3336",
      "description": "Servidor local para desenvolvimento"
    },
    {
      "url": "https://info-api.infobrasilsistemas.com.br",
      "description": "Servidor de produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do usuário",
            "example": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
          },
          "username": {
            "type": "string",
            "description": "Nome de usuário",
            "example": "infomobile"
          },
          "permissions": {
            "description": "Lista de permissões permitidas",
            "example": [
              "tenant.orders.view",
              "tenant.products.view"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roles": {
            "description": "Lista de papéis/roles do usuário",
            "example": [
              "admin",
              "client"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "username",
          "permissions",
          "roles"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do usuário",
            "example": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
          },
          "username": {
            "type": "string",
            "description": "Nome de usuário",
            "example": "infomobile"
          },
          "role": {
            "type": "object",
            "description": "Função/Role do usuário",
            "example": "client",
            "nullable": true
          },
          "permissions": {
            "description": "Lista de permissões permitidas",
            "example": [
              "tenant.orders.view",
              "tenant.products.view"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "username",
          "role",
          "permissions"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Token de acesso JWT",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refresh_token": {
            "type": "string",
            "description": "Token de refresh para obter novos tokens de acesso",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "description": "Dados do usuário autenticado",
            "allOf": [
              {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            ]
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Token de refresh para obter um novo token de acesso"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Novo token de acesso JWT",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "access_token"
        ]
      },
...
      "PaymentMethodResponseDto": {
        "type": "object",
        "properties": {
          "FPG_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da forma de pagamento"
          },
          "FPG_DESCRICAO": {
            "type": "string",
            "example": "Dinheiro",
            "description": "Descrição da forma de pagamento"
          }
        },
        "required": [
          "FPG_CODIGO",
          "FPG_DESCRICAO"
        ]
      },
      "PlanLimitsDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do plano",
            "example": "Bronze"
          },
          "reqMin": {
            "type": "number",
            "description": "Limite de requisições por minuto",
            "example": 100
          },
          "reqMonth": {
            "type": "number",
            "description": "Limite de requisições por mês",
            "example": 15000
          },
          "maxPageSize": {
            "type": "number",
            "description": "Tamanho máximo de página permitido",
            "example": 50
          },
          "maxDateRangeDays": {
            "type": "number",
            "description": "Intervalo máximo de dias para filtros de data",
            "example": 30
          }
        },
        "required": [
          "name",
          "reqMin",
          "reqMonth",
          "maxPageSize",
          "maxDateRangeDays"
        ]
      },
      "PlanUsageDto": {
        "type": "object",
        "properties": {
          "reqsMinute": {
            "type": "number",
            "description": "Requisições feitas no último minuto",
            "example": 5
          },
          "reqsMonth": {
            "type": "number",
            "description": "Requisições feitas no mês corrente",
            "example": 1200
          },
          "minutePercentage": {
            "type": "number",
            "description": "Porcentagem de uso do limite por minuto",
            "example": 5
          },
          "monthPercentage": {
            "type": "number",
            "description": "Porcentagem de uso do limite mensal",
            "example": 8
          }
        },
        "required": [
          "reqsMinute",
          "reqsMonth",
          "minutePercentage",
          "monthPercentage"
        ]
      },
      "UserStatsResponseDto": {
        "type": "object",
        "properties": {
          "limits": {
            "description": "Limites permitidos no plano do usuário",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanLimitsDto"
              }
            ]
          },
          "usage": {
            "description": "Métricas de consumo e utilização",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanUsageDto"
              }
            ]
          }
        },
        "required": [
          "limits",
          "usage"
        ]
      },
      "PlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do plano",
            "example": "d3b07384-d113-49c5-a506-6997a9f60bc8"
          },
          "name": {
            "type": "string",
            "description": "Nome do plano",
            "example": "Bronze"
          },
          "reqMin": {
            "type": "number",
            "description": "Limite de requisições por minuto",
            "example": 100
          },
          "reqMonth": {
            "type": "number",
            "description": "Limite de requisições por mês",
            "example": 15000
          },
          "maxPageSize": {
            "type": "number",
            "description": "Tamanho máximo de página permitido",
            "example": 50
          },
          "maxDateRangeDays": {
            "type": "number",
            "description": "Intervalo máximo de dias para filtros de data",
            "example": 30
          }
        },
        "required": [
          "id",
          "name",
          "reqMin",
          "reqMonth",
          "maxPageSize",
          "maxDateRangeDays"
        ]
      },
...
      "EmployeeResponseDto": {
        "type": "object",
        "properties": {
          "FUN_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código do funcionário"
          },
          "FCA_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da função"
          },
          "FCA_NOME": {
            "type": "string",
            "example": "Vendedor",
            "description": "Descrição da função (Cargo)"
          },
          "FUN_NOME": {
            "type": "string",
            "example": "João da Silva",
            "description": "Nome do funcionário"
          },
          "FUN_APELIDO": {
            "type": "string",
            "example": "João",
            "description": "Apelido do funcionário"
          },
          "FUN_FONE": {
            "type": "string",
            "example": "1133334444",
            "description": "Telefone"
          },
          "FUN_CELULAR": {
            "type": "string",
            "example": "11999998888",
            "description": "Celular"
          },
          "FUN_EMAIL": {
            "type": "string",
            "example": "joao@email.com",
            "description": "E-mail"
          },
          "FUN_SITUACAO": {
            "type": "string",
            "example": "A",
            "description": "Situação (A = Ativo, I = Inativo)"
          },
          "LOJ_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da loja associada"
          }
        },
        "required": [
          "FUN_CODIGO",
          "FCA_CODIGO",
          "FCA_NOME",
          "FUN_NOME",
          "FUN_APELIDO",
          "FUN_FONE",
          "FUN_CELULAR",
          "FUN_EMAIL",
          "FUN_SITUACAO",
          "LOJ_CODIGO"
        ]
      },
      "EmployeeDetailResponseDto": {
        "type": "object",
        "properties": {
          "FUN_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código do funcionário"
          },
          "FCA_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da função"
          },
          "FCA_NOME": {
            "type": "string",
            "example": "Vendedor",
            "description": "Descrição da função (Cargo)"
          },
          "FUN_NOME": {
            "type": "string",
            "example": "João da Silva",
            "description": "Nome do funcionário"
          },
          "FUN_APELIDO": {
            "type": "string",
            "example": "João",
            "description": "Apelido do funcionário"
          },
          "FUN_FONE": {
            "type": "string",
            "example": "1133334444",
            "description": "Telefone"
          },
          "FUN_CELULAR": {
            "type": "string",
            "example": "11999998888",
            "description": "Celular"
          },
          "FUN_EMAIL": {
            "type": "string",
            "example": "joao@email.com",
            "description": "E-mail"
          },
          "FUN_SITUACAO": {
            "type": "string",
            "example": "A",
            "description": "Situação (A = Ativo, I = Inativo)"
          },
          "LOJ_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da loja associada"
          },
          "FUN_ENDERECO": {
            "type": "string",
            "example": "Rua das Flores, 123",
            "description": "Endereço do funcionário"
          },
          "FUN_BAIRRO": {
            "type": "string",
            "example": "Centro",
            "description": "Bairro"
          },
          "FUN_CIDADE": {
            "type": "string",
            "example": "São Paulo",
            "description": "Cidade"
          },
          "FUN_UF": {
            "type": "string",
            "example": "SP",
            "description": "UF"
          },
          "FUN_CEP": {
            "type": "string",
            "example": "01000-000",
            "description": "CEP"
          },
          "FUN_CPF": {
            "type": "string",
            "example": "123.456.789-00",
            "description": "CPF"
          },
          "FUN_IDENTIDADE": {
            "type": "string",
            "example": "12.345.678-9",
            "description": "Identidade (RG)"
          },
          "FUN_DATANASC": {
            "format": "date-time",
            "type": "string",
            "example": "1990-01-01",
            "description": "Data de nascimento"
          },
          "FUN_DATAADMISSAO": {
            "format": "date-time",
            "type": "string",
            "example": "2020-01-01",
            "description": "Data de admissão"
          },
          "FUN_DATADEMISSAO": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "Data de demissão"
          }
        },
        "required": [
          "FUN_CODIGO",
          "FCA_CODIGO",
          "FCA_NOME",
          "FUN_NOME",
          "FUN_APELIDO",
          "FUN_FONE",
          "FUN_CELULAR",
          "FUN_EMAIL",
          "FUN_SITUACAO",
          "LOJ_CODIGO",
          "FUN_ENDERECO",
          "FUN_BAIRRO",
          "FUN_CIDADE",
          "FUN_UF",
          "FUN_CEP",
          "FUN_CPF",
          "FUN_IDENTIDADE",
          "FUN_DATANASC",
          "FUN_DATAADMISSAO",
          "FUN_DATADEMISSAO"
        ]
      },
...
    }
  }
}